refactor(termiod): move the foreground and the wire encoders out of session.rs
Neither depends on a session. `session::wire` turns what the VT engine
holds into what the protocol carries — a cell, a damage set, a scrollback
capture — and is a function of its arguments, which is why its test can
move with it and stop needing a session to run.

`session::foreground` answers "which agent is in there". Its three fields
were loose on `Session`; together they are one thing, with `pending`
guarding the dispatch that fills the sample and the pinned executable
arriving from the same resolution. The split between the cheap half on the
actor and the process-table walk off it is now stated where it lives.

`Session` goes from 24 fields to 22.
